+++ server/db/update_scripts/32_update_v6.17.0_v7.0.0.sql
@@ -728,3 +728,12 @@ INSERT IGNORE INTO `fields` (`name`, `id_type`, `display`) VALUES ('height', get
 INSERT IGNORE INTO `fields` (`name`, `id_type`, `display`) VALUES ('width', get_field_type_id('text'), '0');
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('image'), get_field_id('height'), '', 'Specify the height of the image. The height can be set in pixels (e.g., 200), as a percentage of its parent container (e.g., 50%), or set to `auto` to maintain the aspect ratio based on the width.');
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('image'), get_field_id('width'), '', 'Specify the width of the image. The width can be set in pixels (e.g., 200), as a percentage of its parent container (e.g., 50%), or set to `auto` to maintain the aspect ratio based on the height.');
+
+-- add cache page
+INSERT IGNORE INTO `pages` (`id`, `keyword`, `url`, `protocol`, `id_actions`, `id_navigation_section`, `parent`, `is_headless`, `nav_position`, `footer_position`, `id_type`, `id_pageAccessTypes`) 
+VALUES (NULL, 'cache', '/admin/cache', 'GET|POST', '0000000002', NULL, '0000000009', '0', 1001, NULL, '0000000001', (SELECT id FROM lookups WHERE type_code = "pageAccessTypes" AND lookup_code = "web"));
+SET @id_page = (SELECT id FROM pages WHERE keyword = 'cache');
+
+INSERT IGNORE INTO `pages_fields_translation` (`id_pages`, `id_fields`, `id_languages`, `content`) VALUES (@id_page, get_field_id('label'), '0000000002', 'Cache');
+INSERT IGNORE INTO `pages_fields_translation` (`id_pages`, `id_fields`, `id_languages`, `content`) VALUES (@id_page, get_field_id('title'), '0000000002', 'Cache');
+INSERT IGNORE INTO `acl_groups` (`id_groups`, `id_pages`, `acl_select`, `acl_insert`, `acl_update`, `acl_delete`) VALUES ('0000000001', @id_page, '1', '0', '1', '0');